+++ packages/gas-report/ts/common.ts
@@ -1,0 +1,16 @@
+export type GasReportEntry = {
+  file: string;
+  test: string;
+  name: string;
+  gasUsed: number;
+  prevGasUsed?: number;
+};
+
+export type GasReport = GasReportEntry[];
+
+export type CommandOptions = {
+  path: string[];
+  save?: string;
+  compare?: string;
+  stdin?: boolean;
+};
